This is a squashed commit that ignores all previous history.
All the relevant code in this commit is a complete rewrite of previous work and the history of developing this new code is not useful but it is still maintained here: https://github.com/ShahimEssaid/comp-loinc/tree/dev-se in case it might be useful.  Also see the main-old branch for one previous PR that is no longer applicable to the new code, and was left out of this new "main" history.

There will likely be more Git history rewrites to remove large files that were committed in the past and would be good to remove to reduce the overall size of the repository. This will have to rewrite the history of "main" and have it force pushed but it will not likely affect anything you do that is based on the new implementation. It will only remove non-code files from the history.